Add manifest blob
