+++ radix-engine-interface/src/data/types/mod.rs
@@ -1,11 +1,11 @@
-mod blob;
+mod manifest_blob;
 mod manifest_bucket;
 mod manifest_expression;
 mod manifest_proof;
 mod ownership;
 mod reference;
 
-pub use blob::*;
+pub use manifest_blob::*;
 pub use manifest_bucket::*;
 pub use manifest_expression::*;
 pub use manifest_proof::*;
